+++ examples/code-gen-all.sh
@@ -44,10 +44,10 @@ update_and_run_project() {
     set_project_to_use_nuget
 }
 
+update_and_run_project "$EXAMPLES_DIR/Tutorial1-complete/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/LaserTagMenu1/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/BlankTemplate/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/Blinky1/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/Blinky1Printf/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/ButtonSm1Cpp/CodeGen/"
 update_and_run_project "$EXAMPLES_DIR/Tutorial1-blank/CodeGen/"
-update_and_run_project "$EXAMPLES_DIR/Tutorial1-complete/CodeGen/"