+++ src/StateSmith/Input/DrawIo/MxCell.cs
@@ -1,0 +1,19 @@
+﻿namespace StateSmith.Input.DrawIo;
+#nullable enable
+
+public class MxCell
+{
+    public enum Type { None, Edge, Vertex }
+
+    public string id;
+    public string? parent;
+    public string? label;
+    public Type type = Type.None;
+    public string? source;
+    public string? target;
+
+    public MxCell(string id)
+    {
+        this.id = id;
+    }
+}